+++ src/main.ts
@@ -1,9 +1,10 @@
-import { useAsyncData, unFirstArgumentEnhanced } from "./use-async-data"
+import { useAsyncData, getAsyncDataContext, unFirstArgumentEnhanced } from "./use-async-data"
 import { useAsync, useAsyncFunction } from "./use-async"
 
 export { 
   useAsync, 
   useAsyncFunction,
   useAsyncData, 
-  unFirstArgumentEnhanced
+  getAsyncDataContext,
+  unFirstArgumentEnhanced,
 }